change: use pykunai

+++ src/kunai_sandbox/__init__.py
@@ -11,126 +11,13 @@ import shlex
 import sys
 import shutil
 import tempfile
-import re
 import json
 import hashlib
 import weakref
 
+from pykunai.event import Query
 from qemu.qmp import QMPClient
 from datetime import datetime, timezone
-
-
-class Query(object):
-    def __init__(self, recurse=True):
-        self.guids = set()
-        self.hashes = set()
-        self.regexes = {}
-        # track child tasks
-        self.recurse = recurse
-        self.filter_in = set()
-        self.filter_out = set()
-
-    def add_guids(self, guids):
-        for guid in guids:
-            # guid normalization
-            guid = guid.strip("{}")
-            self.guids.add(guid)
-
-    def add_hashes(self, hashes: list):
-        for h in hashes:
-            self.hashes.add(h)
-
-    def add_regexp(self, regexes):
-        for regex in regexes:
-            self.regexes[regex] = re.compile(regex, re.I)
-
-    def add_filters(self, filters):
-        for f in filters:
-            f = int(f)
-            if f > 0:
-                self.filter_in.add(f)
-            else:
-                self.filter_out.add(-f)
-
-    def _update(self, event):
-        """
-        Update query object from an event
-        """
-        if self.recurse:
-            if "info" in event:
-                if "task" in event["info"]:
-                    guid = event["info"]["task"]["guuid"]
-                    self.add_guids([guid])
-
-    def _match_regex(self, s):
-        s = str(s)
-        for k, r in self.regexes.items():
-            if r.search(s):
-                return True
-        return False
-
-    def _filtered_out(self, eventid):
-        if len(self.filter_in):
-            if eventid not in self.filter_in:
-                return True
-        if len(self.filter_out):
-            if eventid in self.filter_out:
-                return True
-        return False
-
-    def _filtered_in(self, eventid):
-        if len(self.filter_in):
-            if eventid in self.filter_in:
-                return True
-        if len(self.filter_out):
-            if eventid not in self.filter_out:
-                return True
-        return False
-
-    def _recursive_walk(self, dictionary):
-        for key, value in dictionary.items():
-            if isinstance(value, dict):
-                for k, v in self._recursive_walk(value):
-                    yield k, v
-            yield key, value
-
-    def match(self, event):
-        if "info" in event and "data" in event:
-            eventdata = event["data"]
-            task_info = event["info"]["task"]
-            ptask_info = event["info"]["parent_task"]
-
-            # check for event id
-            if "event" in event["info"]:
-                if "id" in event["info"]["event"]:
-                    eid = int(event["info"]["event"]["id"])
-                    if self._filtered_out(eid):
-                        return False
-                    if self._filtered_in(eid):
-                        return True
-
-            # check for event data
-            if "data" in event:
-                # if create process
-                if (
-                    task_info["guuid"] in self.guids
-                    or ptask_info["guuid"] in self.guids
-                ):
-                    self._update(event)
-                    return True
-
-                for k, v in self._recursive_walk(event):
-                    # check for Hashes
-                    if k in ["md5", "sha1", "sha256", "sha512"]:
-                        if v in self.hashes:
-                            self._update(event)
-                            return True
-
-                    if self._match_regex(v):
-                        self._update(event)
-                        return True
-
-        return False
 
 
 class Sandbox:
@@ -146,7 +33,8 @@ class Sandbox:
         weakref.finalize(self, self.cleanup)
 
     def cleanup(self):
-        os.remove(self._pcap_file)
+        if os.path.isfile(self._pcap_file):
+            os.remove(self._pcap_file)
 
     @property
     def pcap_file(self) -> str:
@@ -342,7 +230,7 @@ class Sandbox:
 
     def run_qemu_command(self, command):
         if self._qemu_process.poll() is not None:
-            raise Exception(f"qemu command stopped unexpectedely")
+            raise Exception("qemu command stopped unexpectedely")
 
         async def _run_qemu_cmd(cmd):
             qmp = QMPClient("blah")
@@ -364,7 +252,7 @@ class Sandbox:
 def sandbox_stop_no_fail(sbx):
     try:
         sbx.stop()
-    except:
+    except Exception:
         pass
